requirements.txt
# Зависимости проекта
# Telegram Bot
python-telegram-bot>=20.0

# Работа с Excel
openpyxl>=3.0.0
pandas>=2.0.0

# Работа с датами
python-dateutil>=2.8.0

# Переменные окружения
python-dotenv>=1.0.0

# Валидация данных
pydantic>=2.0.0
